Finished tests for id company fetching
