+++ test/end-to-end/company.js
@@ -13,19 +13,9 @@ import hash from "../../src/lib/passwordHashing";
 import Account from "../../src/models/Account";
 import Company from "../../src/models/Company";
 import Offer from "../../src/models/Offer";
-import CompanyConstants from "../../src/models/constants/Company";
 import OfferConstants from "../../src/models/constants/Offer";
 import withGodToken from "../utils/GodToken";
 import { DAY_TO_MS } from "../utils/TimeConstants";
-
-const getCompanies = async (options) =>
-    [...(await Company.find(options)
-        .sort({ name: "asc" }) // sort them to match what gets returned by the service
-        .exec())]
-        .map((company) => ({
-            ...company.toObject(),
-            _id: company._id.toString(),
-        }));
 
 describe("Company endpoint", () => {
 
@@ -53,752 +43,6 @@ describe("Company endpoint", () => {
         contacts: ["112", "122"],
         hasFinishedRegistration: true,
         ...params,
-    });
-
-
-    describe("GET /company", () => {
-        beforeAll(async () => {
-            await Company.deleteMany({});
-        });
-
-        test("should return an empty array", async () => {
-            const res = await request()
-                .get("/company");
-            expect(res.status).toBe(HTTPStatus.OK);
-            expect(res.body.companies).toEqual([]);
-            expect(res.body.totalDocCount).toEqual(0);
-        });
-
-        test("should return the newly created company", async () => {
-            await Company.create({ name: "Company" });
-            const res = await request()
-                .get("/company");
-            const companies = await getCompanies({});
-            expect(res.status).toBe(HTTPStatus.OK);
-            expect(res.body.companies).toEqual(companies);
-            expect(res.body.totalDocCount).toEqual(1);
-        });
-
-        test("should not return blocked created company if not logged in", async () => {
-            await Company.deleteMany({});
-            await Company.create({ name: "Company", isBlocked: true });
-            const res = await request()
-                .get("/company");
-            expect(res.status).toBe(HTTPStatus.OK);
-            expect(res.body.companies).toEqual([]);
-            expect(res.body.totalDocCount).toEqual(0);
-        });
-
-
-        describe("With Auth", () => {
-            const test_agent = agent();
-
-            const test_user_admin = {
-                email: "admin@email.com",
-                password: "password123",
-            };
-
-            beforeEach(async () => {
-                await Company.deleteMany({});
-                await Account.deleteMany({});
-                await Account.create({
-                    email: test_user_admin.email,
-                    password: await hash(test_user_admin.password),
-                    isAdmin: true
-                });
-            });
-
-            test("should return blocked created company logged in as admin", async () => {
-                await Company.create({ name: "Company", isBlocked: true });
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_admin)
-                    .expect(200);
-
-                const res = await test_agent
-                    .get("/company");
-                const companies = await getCompanies({});
-                expect(res.status).toBe(HTTPStatus.OK);
-                expect(res.body.companies).toEqual(companies);
-                expect(res.body.totalDocCount).toEqual(1);
-            });
-
-        });
-
-        describe("Disabled companies", () => {
-            let test_company, disabled_test_company;
-            const test_user_admin = {
-                email: "admin@email.com",
-                password: "password123",
-            };
-            const test_user_company = {
-                email: "company@email.com",
-                password: "password123",
-            };
-
-            const test_agent = agent();
-
-            beforeAll(async () => {
-
-                await test_agent
-                    .delete("/auth/login")
-                    .expect(HTTPStatus.OK);
-
-                await Company.deleteMany({});
-
-                test_company = {
-                    name: "test-company"
-                };
-
-                disabled_test_company = {
-                    name: "disabled-test-company",
-                    isDisabled: true
-                };
-
-                const companies = await Company.create([test_company, disabled_test_company]);
-
-                await Account.deleteMany({});
-                await Account.create({
-                    email: test_user_admin.email,
-                    password: await hash(test_user_admin.password),
-                    isAdmin: true
-                });
-                await Account.create({
-                    email: test_user_company.email,
-                    password: await hash(test_user_company.password),
-                    company: companies[0]._id
-                });
-            });
-
-            afterEach(async () => {
-                await test_agent
-                    .delete("/auth/login")
-                    .expect(HTTPStatus.OK);
-            });
-
-            test("should return both companies if god token is sent", async () => {
-
-                const res = await test_agent
-                    .get("/company")
-                    .send(withGodToken());
-
-                const companies = await getCompanies({});
-
-                expect(res.status).toBe(HTTPStatus.OK);
-                expect(res.body.companies).toEqual(companies);
-                expect(res.body.totalDocCount).toEqual(2);
-
-            });
-
-            test("should return both companies if logged as admin", async () => {
-
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_admin)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get("/company");
-
-                const companies = await getCompanies({});
-                expect(res.status).toBe(HTTPStatus.OK);
-                expect(res.body.companies).toEqual(companies);
-                expect(res.body.totalDocCount).toEqual(2);
-
-            });
-
-            test("should return only the enabled company if logged as unprivileged user", async () => {
-
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_company)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get("/company");
-
-                const companies = await getCompanies({ isDisabled: { $ne: true } });
-                expect(res.status).toBe(HTTPStatus.OK);
-                expect(res.body.companies).toEqual(companies);
-                expect(res.body.totalDocCount).toEqual(1);
-            });
-
-            test("should return only the enabled company if not logged", async () => {
-
-                const res = await test_agent
-                    .get("/company");
-
-                const companies = await getCompanies({ isDisabled: { $ne: true } });
-                expect(res.status).toBe(HTTPStatus.OK);
-                expect(res.body.companies).toEqual(companies);
-                expect(res.body.totalDocCount).toEqual(1);
-            });
-        });
-    });
-
-    describe("GET /company/:companyId", () => {
-        let test_company_without_offers,
-            test_company_with_offers_below_limit,
-            test_company_with_offers_at_limit,
-            test_company_with_offers_above_limit,
-            test_company_with_hidden_offer,
-            test_offer_hidden,
-            test_company_disabled,
-            test_company_blocked,
-            test_company_with_unfinished_registration;
-
-        const test_user_with_unfinished_registration = {
-            email: "unfinished@email.com",
-            password: "password123",
-        };
-        const test_user_with_hidden_offer = {
-            email: "hidden@email.com",
-            password: "password123",
-        };
-        const test_user_admin = {
-            email: "admin@email.com",
-            password: "password123",
-            isAdmin: true,
-        };
-        const test_user_disabled_company = {
-            email: "disabled@email.com",
-            password: "password123",
-        };
-        const test_user_blocked_company = {
-            email: "blocked@email.com",
-            password: "password123",
-        };
-        const test_company_data = {
-            name: "test-company",
-            hasFinishedRegistration: true,
-            logo: "http://awebsite.com/alogo.jpg",
-        };
-
-        const test_agent = agent();
-
-        const createTestOffers = (length, company) =>
-            Promise.all(
-                Array.from({ length }, () =>
-                    Offer.create(
-                        generateTestOffer({
-                            publishDate: new Date(
-                                Date.now() - DAY_TO_MS
-                            ).toISOString(),
-                            publishEndDate: new Date(
-                                Date.now() + DAY_TO_MS
-                            ).toISOString(),
-                            owner: company._id,
-                            ownerName: company.name,
-                            ownerLogo: company.logo,
-                        })
-                    )
-                )
-            );
-
-        beforeAll(async () => {
-            await Company.deleteMany({});
-
-            [
-                test_company_without_offers,
-                test_company_with_offers_below_limit,
-                test_company_with_offers_at_limit,
-                test_company_with_offers_above_limit,
-                test_company_with_hidden_offer,
-                test_company_disabled,
-                test_company_blocked,
-                test_company_with_unfinished_registration,
-            ] = await Company.create([
-                test_company_data,
-                test_company_data,
-                test_company_data,
-                test_company_data,
-                test_company_data,
-                { ...test_company_data, isDisabled: true },
-                { ...test_company_data, isBlocked: true },
-                { ...test_company_data, hasFinishedRegistration: false },
-            ]);
-
-            await Account.deleteMany({});
-
-            await Promise.all([
-                Account.create({
-                    email: test_user_admin.email,
-                    password: await hash(test_user_admin.password),
-                    isAdmin: true,
-                }),
-                Account.create({
-                    email: test_user_with_unfinished_registration.email,
-                    password: await hash(
-                        test_user_with_unfinished_registration.password
-                    ),
-                    company: test_company_with_unfinished_registration._id,
-                }),
-                Account.create({
-                    email: test_user_with_hidden_offer.email,
-                    password: await hash(test_user_with_hidden_offer.password),
-                    company: test_company_with_hidden_offer._id,
-                }),
-                Account.create({
-                    email: test_user_disabled_company.email,
-                    password: await hash(test_user_disabled_company.password),
-                    company: test_company_disabled._id,
-                }),
-                Account.create({
-                    email: test_user_blocked_company.email,
-                    password: await hash(test_user_blocked_company.password),
-                    company: test_company_blocked._id,
-                }),
-            ]);
-
-            test_offer_hidden = await Offer.create(
-                generateTestOffer({
-                    isHidden: true,
-                    owner: test_company_with_hidden_offer._id.toString(),
-                    ownerName: test_company_with_hidden_offer.name,
-                    ownerLogo: test_company_with_hidden_offer.logo,
-                })
-            );
-        });
-
-        afterEach(async () => {
-            await test_agent.delete("/auth/login").expect(HTTPStatus.OK);
-        });
-
-        describe("Without Auth", () => {
-            test("should fail if invalid id", async () => {
-                const res = await test_agent
-                    .get("/company/123")
-                    .expect(HTTPStatus.UNPROCESSABLE_ENTITY);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.VALIDATION_ERROR
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            param: "companyId",
-                            msg: ValidationReasons.OBJECT_ID,
-                        }),
-                    ])
-                );
-            });
-
-            test("should fail if company does not exist", async () => {
-                const id = "111111111111111111111111";
-                const res = await test_agent
-                    .get(`/company/${id}`)
-                    .expect(HTTPStatus.UNPROCESSABLE_ENTITY);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.VALIDATION_ERROR
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            param: "companyId",
-                            msg: ValidationReasons.COMPANY_NOT_FOUND(id),
-                        }),
-                    ])
-                );
-            });
-
-            test("should succeed when the company has no offers", async () => {
-                const res = await test_agent
-                    .get(`/company/${test_company_without_offers.id}`)
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty("offers", []);
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_without_offers._id.toString()
-                );
-            });
-
-            test("should return all offers when below limit", async () => {
-                const offers = await createTestOffers(
-                    CompanyConstants.offers.max_profile_visible - 1,
-                    test_company_with_offers_below_limit
-                );
-
-                const res = await test_agent
-                    .get(`/company/${test_company_with_offers_below_limit._id}`)
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty("offers");
-                expect(res.body.offers).toHaveLength(
-                    CompanyConstants.offers.max_profile_visible - 1
-                );
-                expect(res.body.offers.map((x) => x._id).sort()).toEqual(
-                    offers.map((x) => x._id.toString()).sort()
-                );
-
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_with_offers_below_limit._id.toString()
-                );
-            });
-
-            test("should return all offers when at limit", async () => {
-                const offers = await createTestOffers(
-                    CompanyConstants.offers.max_profile_visible,
-                    test_company_with_offers_at_limit
-                );
-
-                const res = await test_agent
-                    .get(`/company/${test_company_with_offers_at_limit._id}`)
-                    .expect(HTTPStatus.OK);
-                expect(res.body).toHaveProperty("offers");
-                expect(res.body.offers).toHaveLength(
-                    CompanyConstants.offers.max_profile_visible
-                );
-                expect(res.body.offers.map((x) => x._id).sort()).toEqual(
-                    offers.map((x) => x._id.toString()).sort()
-                );
-
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_with_offers_at_limit._id.toString()
-                );
-            });
-
-            test("should limit number of offers", async () => {
-                const offers = await createTestOffers(
-                    CompanyConstants.offers.max_profile_visible + 10,
-                    test_company_with_offers_above_limit
-                );
-
-                const res = await test_agent
-                    .get(`/company/${test_company_with_offers_above_limit._id}`)
-                    .expect(HTTPStatus.OK);
-                expect(res.body).toHaveProperty("offers");
-                expect(res.body.offers).toHaveLength(
-                    CompanyConstants.offers.max_profile_visible
-                );
-                expect(offers.map((x) => x._id.toString())).toEqual(
-                    expect.arrayContaining(res.body.offers.map((x) => x._id))
-                );
-
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_with_offers_above_limit._id.toString()
-                );
-            });
-
-            test("should not return hidden offers", async () => {
-                const res = await test_agent
-                    .get(`/company/${test_company_with_hidden_offer._id}`)
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty("offers", []);
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_with_hidden_offer._id.toString()
-                );
-            });
-
-            test("should fail if company is disabled", async () => {
-                const res = await test_agent
-                    .get(`/company/${test_company_disabled._id}`)
-                    .expect(HTTPStatus.UNPROCESSABLE_ENTITY);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.VALIDATION_ERROR
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            param: "companyId",
-                            msg: ValidationReasons.COMPANY_NOT_FOUND(
-                                test_company_disabled._id
-                            ),
-                        }),
-                    ])
-                );
-            });
-
-            test("should fail if company is blocked", async () => {
-                const res = await test_agent
-                    .get(`/company/${test_company_blocked._id}`)
-                    .expect(HTTPStatus.UNPROCESSABLE_ENTITY);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.VALIDATION_ERROR
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            param: "companyId",
-                            msg: ValidationReasons.COMPANY_NOT_FOUND(
-                                test_company_blocked._id
-                            ),
-                        }),
-                    ])
-                );
-            });
-
-            test("should fail if company hasn't finished registration", async () => {
-                const res = await test_agent
-                    .get(
-                        `/company/${test_company_with_unfinished_registration._id}`
-                    )
-                    .expect(HTTPStatus.UNPROCESSABLE_ENTITY);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.VALIDATION_ERROR
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            param: "companyId",
-                            msg: ValidationReasons.COMPANY_NOT_FOUND(
-                                test_company_with_unfinished_registration._id
-                            ),
-                        }),
-                    ])
-                );
-            });
-        });
-
-        describe("With Auth", () => {
-            test("should return hidden offers when user is owner", async () => {
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_with_hidden_offer)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get(`/company/${test_company_with_hidden_offer._id}`)
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty("offers", [
-                    expect.objectContaining({
-                        _id: test_offer_hidden._id.toString(),
-                    }),
-                ]);
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_with_hidden_offer._id.toString()
-                );
-            });
-
-            test("should succeed if company is disabled and user is owner", async () => {
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_disabled_company)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get(`/company/${test_company_disabled._id}`)
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_disabled._id.toString()
-                );
-            });
-
-            test("should fail if company is blocked and user is owner", async () => {
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_blocked_company)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get(`/company/${test_company_blocked._id}`)
-                    .expect(HTTPStatus.FORBIDDEN);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.FORBIDDEN
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            msg: ValidationReasons.COMPANY_BLOCKED,
-                        }),
-                    ])
-                );
-            });
-
-            test("should fail if company hasn't finished registration and user is owner", async () => {
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_with_unfinished_registration)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get(
-                        `/company/${test_company_with_unfinished_registration._id}`
-                    )
-                    .expect(HTTPStatus.FORBIDDEN);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.FORBIDDEN
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            msg: ValidationReasons.REGISTRATION_NOT_FINISHED,
-                        }),
-                    ])
-                );
-            });
-
-            test("should return hidden offers when user is admin", async () => {
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_admin)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get(`/company/${test_company_with_hidden_offer._id}`)
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty("offers", [
-                    expect.objectContaining({
-                        _id: test_offer_hidden._id.toString(),
-                    }),
-                ]);
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_with_hidden_offer._id.toString()
-                );
-            });
-
-            test("should succeed if company is disabled and user is admin", async () => {
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_admin)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get(`/company/${test_company_disabled._id}`)
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_disabled._id.toString()
-                );
-            });
-
-            test("should succeed if company is blocked and user is admin", async () => {
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_admin)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get(`/company/${test_company_blocked._id}`)
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_blocked._id.toString()
-                );
-            });
-
-            test("should fail if company hasn't finished registration and user is admin", async () => {
-                await test_agent
-                    .post("/auth/login")
-                    .send(test_user_admin)
-                    .expect(HTTPStatus.OK);
-
-                const res = await test_agent
-                    .get(
-                        `/company/${test_company_with_unfinished_registration._id}`
-                    )
-                    .expect(HTTPStatus.FORBIDDEN);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.FORBIDDEN
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            msg: ValidationReasons.REGISTRATION_NOT_FINISHED,
-                        }),
-                    ])
-                );
-            });
-
-            test("should return hidden offers when user is god", async () => {
-                const res = await test_agent
-                    .get(`/company/${test_company_with_hidden_offer._id}`)
-                    .send(withGodToken())
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty("offers", [
-                    expect.objectContaining({
-                        _id: test_offer_hidden._id.toString(),
-                    }),
-                ]);
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_with_hidden_offer._id.toString()
-                );
-            });
-
-            test("should succeed if company is disabled and user is god", async () => {
-                const res = await test_agent
-                    .get(`/company/${test_company_disabled._id}`)
-                    .send(withGodToken())
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_disabled._id.toString()
-                );
-            });
-
-            test("should succeed if company is blocked and user is god", async () => {
-                const res = await test_agent
-                    .get(`/company/${test_company_blocked._id}`)
-                    .send(withGodToken())
-                    .expect(HTTPStatus.OK);
-
-                expect(res.body).toHaveProperty(
-                    "company._id",
-                    test_company_blocked._id.toString()
-                );
-            });
-
-            test("should fail if company hasn't finished registration and user is god", async () => {
-                const res = await test_agent
-                    .get(
-                        `/company/${test_company_with_unfinished_registration._id}`
-                    )
-                    .send(withGodToken())
-                    .expect(HTTPStatus.FORBIDDEN);
-
-                expect(res.body).toHaveProperty(
-                    "error_code",
-                    ErrorTypes.FORBIDDEN
-                );
-                expect(res.body).toHaveProperty(
-                    "errors",
-                    expect.arrayContaining([
-                        expect.objectContaining({
-                            msg: ValidationReasons.REGISTRATION_NOT_FINISHED,
-                        }),
-                    ])
-                );
-            });
-        });
     });
 
     describe("PUT /company/:companyId/block", () => {
